+++ modules/container-web-app/variables.tf
@@ -291,3 +291,15 @@ variable "app_service_logs" {
   })
   default = null
 }
+
+variable "identity" {
+  description = "Map with identity block information."
+  type = object({
+    type         = string
+    identity_ids = list(string)
+  })
+  default = {
+    type         = "SystemAssigned"
+    identity_ids = []
+  }
+}