+++ modules/linux-web-app/r-appservice.tf
@@ -126,8 +126,12 @@ resource "azurerm_linux_web_app" "app_service_linux" {
   client_certificate_enabled = var.client_certificate_enabled
   https_only                 = var.https_only
 
-  identity {
-    type = "SystemAssigned"
+  dynamic "identity" {
+    for_each = var.identity[*]
+    content {
+      type         = var.identity.type
+      identity_ids = var.identity.identity_ids
+    }
   }
 
   dynamic "backup" {
@@ -327,8 +331,12 @@ resource "azurerm_linux_web_app_slot" "app_service_linux_slot" {
   client_affinity_enabled = var.client_affinity_enabled
   https_only              = var.https_only
 
-  identity {
-    type = "SystemAssigned"
+  dynamic "identity" {
+    for_each = var.identity[*]
+    content {
+      type         = var.identity.type
+      identity_ids = var.identity.identity_ids
+    }
   }
 
   dynamic "storage_account" {